third_party/WebKit/LayoutTests/fast/mediastream/RTCPeerConnection-expected.txt

Tests the RTCPeerConnection constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new webkitRTCPeerConnection(null); did not throw exception.
PASS new webkitRTCPeerConnection(null, null); did not throw exception.
PASS new webkitRTCPeerConnection(undefined); did not throw exception.
PASS new webkitRTCPeerConnection(undefined, undefined); did not throw exception.
PASS new webkitRTCPeerConnection(); threw exception TypeError: Failed to construct 'RTCPeerConnection': 1 argument required, but only 0 present..
PASS new webkitRTCPeerConnection(''); threw exception TypeError: Failed to construct 'RTCPeerConnection': parameter 1 ('rtcConfiguration') is not an object..
PASS new webkitRTCPeerConnection(null, ''); threw exception TypeError: Failed to construct 'RTCPeerConnection': parameter 2 ('mediaConstraints') is not an object..
PASS new webkitRTCPeerConnection({iceServers:[]}, null); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{url:'stun:foo.com'}]}, null); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{url:'turn:foo.com', credential:'x'}]}, null); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{url:'turn:foo.com', credential:'x'},{url:'stun:bar.com'}]}, null); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}]}, null); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[{urls:['stun:foo.com', 'turn:foo.com']}]}, null); did not throw exception.
PASS new webkitRTCPeerConnection({fooServers:[]}, null); threw exception TypeError: Failed to construct 'RTCPeerConnection': Malformed RTCConfiguration.
PASS new webkitRTCPeerConnection({iceServers:true}, null); threw exception TypeError: Failed to construct 'RTCPeerConnection': Malformed RTCConfiguration.
PASS new webkitRTCPeerConnection({iceServers:[1, 2, 3]}, null); threw exception TypeError: Failed to construct 'RTCPeerConnection': Malformed RTCIceServer.
PASS new webkitRTCPeerConnection({iceServers:[{}]}, null); threw exception TypeError: Failed to construct 'RTCPeerConnection': Malformed RTCIceServer.
PASS new webkitRTCPeerConnection({iceServers:[{url:'foo'}]}, null); threw exception TypeError: Failed to construct 'RTCPeerConnection': Malformed URL.
PASS new webkitRTCPeerConnection({iceServers:[{urls:[1, 'turn:foo.com']}]}, null); threw exception TypeError: Failed to construct 'RTCPeerConnection': Malformed URL.
PASS new webkitRTCPeerConnection({iceServers:[], iceTransports:'none'}, null); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[], iceTransports:'relay'}, null); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[], iceTransports:'all'}, null); did not throw exception.
PASS new webkitRTCPeerConnection({iceServers:[], iceTransports:'foo'}, null); threw exception TypeError: Failed to construct 'RTCPeerConnection': Malformed RTCIceTransports.
PASS new webkitRTCPeerConnection(null, {mandatory:{valid_and_supported_1:1}}); did not throw exception.
PASS new webkitRTCPeerConnection(null, {mandatory:{valid_and_supported_1:1, valid_and_supported_2:1}}); did not throw exception.
PASS new webkitRTCPeerConnection(null, {optional:[{valid_and_supported_1:0}]}); did not throw exception.
PASS new webkitRTCPeerConnection(null, {optional:[{valid_and_supported_1:0},{valid_and_supported_2:0}]}); did not throw exception.
PASS new webkitRTCPeerConnection(null, {optional:[{valid_but_unsupported_1:0},{valid_but_unsupported_2:0}]}); did not throw exception.
PASS new webkitRTCPeerConnection(null, {mandatory:{valid_and_supported_1:66}}); threw exception NotSupportedError: Failed to construct 'RTCPeerConnection': Failed to initialize native PeerConnection..
PASS new webkitRTCPeerConnection(null, {mandatory:{invalid:1}}); threw exception NotSupportedError: Failed to construct 'RTCPeerConnection': Failed to initialize native PeerConnection..
PASS new webkitRTCPeerConnection(null, {mandatory:{valid_but_unsupported_1:1}}); threw exception NotSupportedError: Failed to construct 'RTCPeerConnection': Failed to initialize native PeerConnection..
PASS new webkitRTCPeerConnection(null, {mandatory:{valid_but_unsupported_1:1, valid_and_supported_1:1}}); threw exception NotSupportedError: Failed to construct 'RTCPeerConnection': Failed to initialize native PeerConnection..
PASS new webkitRTCPeerConnection(null, {optional:{valid_and_supported_1:0}}); threw exception TypeError: Failed to construct 'RTCPeerConnection': Malformed constraints object..
PASS new webkitRTCPeerConnection(null, {optional:[{valid_and_supported_1:0,valid_and_supported_2:0}]}); threw exception TypeError: Failed to construct 'RTCPeerConnection': Malformed constraints object..
PASS new webkitRTCPeerConnection(null, {optional:[{invalid:0}]}); threw exception NotSupportedError: Failed to construct 'RTCPeerConnection': Failed to initialize native PeerConnection..
PASS new webkitRTCPeerConnection(null, {valid_and_supported_1:1}); threw exception TypeError: Failed to construct 'RTCPeerConnection': Malformed constraints object..
PASS new webkitRTCPeerConnection(null, {valid_but_unsupported_1:1}); threw exception TypeError: Failed to construct 'RTCPeerConnection': Malformed constraints object..
PASS new webkitRTCPeerConnection(null, {valid_and_supported_2:1, mandatory:{valid_and_supported_1:1}}); threw exception TypeError: Failed to construct 'RTCPeerConnection': Malformed constraints object..
PASS successfullyParsed is true

TEST COMPLETE
